use 32bit feature id

+++ kernel/feature.c
@@ -18,32 +18,32 @@ int ksu_register_feature_handler(const struct ksu_feature_handler *handler)
     }
 
     if (handler->feature_id >= KSU_FEATURE_MAX) {
-        pr_err("feature: invalid feature_id %llu\n", handler->feature_id);
+        pr_err("feature: invalid feature_id %u\n", handler->feature_id);
         return -EINVAL;
     }
 
     if (!handler->get_handler && !handler->set_handler) {
-        pr_err("feature: no handler provided for feature %llu\n", handler->feature_id);
+        pr_err("feature: no handler provided for feature %u\n", handler->feature_id);
         return -EINVAL;
     }
 
     mutex_lock(&feature_mutex);
 
     if (feature_handlers[handler->feature_id]) {
-        pr_warn("feature: handler for %llu already registered, overwriting\n",
+        pr_warn("feature: handler for %u already registered, overwriting\n",
                 handler->feature_id);
     }
 
     feature_handlers[handler->feature_id] = kmalloc(sizeof(struct ksu_feature_handler), GFP_KERNEL);
     if (!feature_handlers[handler->feature_id]) {
-        pr_err("feature: failed to allocate handler for %llu\n", handler->feature_id);
+        pr_err("feature: failed to allocate handler for %u\n", handler->feature_id);
         ret = -ENOMEM;
         goto out;
     }
 
     memcpy(feature_handlers[handler->feature_id], handler, sizeof(struct ksu_feature_handler));
 
-    pr_info("feature: registered handler for %s (id=%llu)\n",
+    pr_info("feature: registered handler for %s (id=%u)\n",
             handler->name ? handler->name : "unknown", handler->feature_id);
 
 out:
@@ -51,19 +51,19 @@ out:
     return ret;
 }
 
-int ksu_unregister_feature_handler(u64 feature_id)
+int ksu_unregister_feature_handler(u32 feature_id)
 {
     int ret = 0;
 
     if (feature_id >= KSU_FEATURE_MAX) {
-        pr_err("feature: invalid feature_id %llu\n", feature_id);
+        pr_err("feature: invalid feature_id %u\n", feature_id);
         return -EINVAL;
     }
 
     mutex_lock(&feature_mutex);
 
     if (!feature_handlers[feature_id]) {
-        pr_warn("feature: no handler registered for %llu\n", feature_id);
+        pr_warn("feature: no handler registered for %u\n", feature_id);
         ret = -ENOENT;
         goto out;
     }
@@ -71,20 +71,20 @@ int ksu_unregister_feature_handler(u64 feature_id)
     kfree(feature_handlers[feature_id]);
     feature_handlers[feature_id] = NULL;
 
-    pr_info("feature: unregistered handler for id=%llu\n", feature_id);
+    pr_info("feature: unregistered handler for id=%u\n", feature_id);
 
 out:
     mutex_unlock(&feature_mutex);
     return ret;
 }
 
-int ksu_get_feature(u64 feature_id, u64 *value, bool *supported)
+int ksu_get_feature(u32 feature_id, u64 *value, bool *supported)
 {
     int ret = 0;
     struct ksu_feature_handler *handler;
 
     if (feature_id >= KSU_FEATURE_MAX) {
-        pr_err("feature: invalid feature_id %llu\n", feature_id);
+        pr_err("feature: invalid feature_id %u\n", feature_id);
         return -EINVAL;
     }
 
@@ -100,21 +100,21 @@ int ksu_get_feature(u64 feature_id, u64 *value, bool *supported)
     if (!handler) {
         *supported = false;
         *value = 0;
-        pr_debug("feature: feature %llu not supported\n", feature_id);
+        pr_debug("feature: feature %u not supported\n", feature_id);
         goto out;
     }
 
     *supported = true;
 
     if (!handler->get_handler) {
-        pr_warn("feature: no get_handler for feature %llu\n", feature_id);
+        pr_warn("feature: no get_handler for feature %u\n", feature_id);
         ret = -EOPNOTSUPP;
         goto out;
     }
 
     ret = handler->get_handler(value);
     if (ret) {
-        pr_err("feature: get_handler for %llu failed: %d\n", feature_id, ret);
+        pr_err("feature: get_handler for %u failed: %d\n", feature_id, ret);
     }
 
 out:
@@ -122,13 +122,13 @@ out:
     return ret;
 }
 
-int ksu_set_feature(u64 feature_id, u64 value)
+int ksu_set_feature(u32 feature_id, u64 value)
 {
     int ret = 0;
     struct ksu_feature_handler *handler;
 
     if (feature_id >= KSU_FEATURE_MAX) {
-        pr_err("feature: invalid feature_id %llu\n", feature_id);
+        pr_err("feature: invalid feature_id %u\n", feature_id);
         return -EINVAL;
     }
 
@@ -137,20 +137,20 @@ int ksu_set_feature(u64 feature_id, u64 value)
     handler = feature_handlers[feature_id];
 
     if (!handler) {
-        pr_err("feature: feature %llu not registered\n", feature_id);
+        pr_err("feature: feature %u not registered\n", feature_id);
         ret = -EOPNOTSUPP;
         goto out;
     }
 
     if (!handler->set_handler) {
-        pr_warn("feature: no set_handler for feature %llu\n", feature_id);
+        pr_warn("feature: no set_handler for feature %u\n", feature_id);
         ret = -EOPNOTSUPP;
         goto out;
     }
 
     ret = handler->set_handler(value);
     if (ret) {
-        pr_err("feature: set_handler for %llu failed: %d\n", feature_id, ret);
+        pr_err("feature: set_handler for %u failed: %d\n", feature_id, ret);
     }
 
 out:
